scripts/requirements.txt
streamlit==1.24.0
opencv-python==4.8.0
tensorflow==2.12.0
numpy==1.24.3
cryptography==41.0.3
boto3==1.28.17
Pillow==10.0.0
scikit-learn==1.3.0
psycopg2-binary==2.9.6
pandas==2.0.3
opencv-python-headless==4.7.0.72
gunicorn>=20.1.0